requirements.txt
docopt>=0.6.1
logbook>=0.6
numpy>=1.8
pylibtiff>=0.3
